backend/app/requirements.txt
fastapi
uvicorn[standard]
chromadb
httpx
pydantic
python-multipart
jinja2
requests
pdfplumber
python-docx
PyMuPDF
pytesseract
reportlab
langchain-community
langchain-ollama
langchain
langchain-prompty
pypdf
langdetect
openpyxl
langsmith
pyyaml
nltk
numpy==1.26.4
scikit-learn